Deploy to production in CI/CD
We only build the docker release img on demand. The docker image is tricky to use
especially with regards to GUI rendering to the client X server when run on a remote
machine. Instead we now deploy directly to the production server using a shell
executor. This is far simpler and since we use virtualenv, just as clean as the docker
approach.