[ENH] Add ramsis ui utils with form loader mixin
The form loader mixin makes form loading simpler and easier to adapt should the location of .ui change one day (e.g. when they will be installed to a different directory with setup.py). It's implemented as a ramsis specific module (hence under RAMSIS.ui instead of the general ui package RAMSIS.ui.base) since it needs to know the path to the directory with the .ui files. To make this more general we would have to configure the .ui path at run-time before any ui modules are imported which is a bit fragile.