[ENH] Add configurable forecast scenario
Provide a new dialog in order to configure a fully featured scenario
(including well, reservoir geometry and stages). Preconfigure the
dialog using a default scenario. When accepting, the dialog operates
directly on the scenario passed and sets attributes accordingly.
Though, it is up to the dialog's client what's happening next (i.e. if
the scenario is saved/updated etc.).

Still there are some TODOs left over:
  - forecast scenario parameter validation (reservoir geometry,
    injection well plan)