[BUG] Remove one-off (singleshot) task property
The one-off task property seems to be legacy. Both ForecastTask and
PeriodicTask are somehow periodic distinguished exclusively by their
corresponding schedule() method implementation. Meanwhile a
PeriodicTask implements its schedule() method in a way that the next
execution is based on the last execution time, the ForecastTask simply
gathers the next forecast from the project's list of forecasts. The
one-off property was not meaningful at all.