[BUGFIX] Reservoir requires data snapshot
On updating the integration tests, it was found that there was
race-condition behaviour concerning the reservoir serialization. What
happened is the same object would get passed to multiple threads which
serialized the data in place. The serialization included the
tranformation of coordinates from local to wsg84, and this was happening
multiple times on the same data when multiple model runs are configured
for a scenario.
The work around is to make a 'snapshot' of the model run, which is
essentially a copy/clone of the data with only the attributes that are
required copied.

Then the snapshot is thrown away and only the data that is needed to be
updated in the db is updated, rather than the whole model run being
merged (which takes a long time as it thinks that all the hydraulics
have been updated too.)