[BUGFIX] Fix transformation of data in db
The bug was that moving the writing of the snapshot and the plan to a
parallel thread meant that while the data was being written, the data
was also being transformed. The way I have solved this is by doing the
transform on the data when it has been serialized so is in a dict rather
than object.