[ENH] Hazard curves and testing correct
The capabilities for oq hazard curves results and the tests have been
added.
There is a strange bug that occurs maybe 1/50 times where I think that a
thread does not finish on time so that the hazard stage, scenaro and
forecast gain a status of ERROR. I am not sure of the origin of this
problem, so leaving a print statement in the state handlers file.