[BUG] Fix for non-updating seismicity models
There was a bug which occured due to the joining properties of the
sqlalchemy tables. Changing from lazy to joined (non automatic join) meant that each of the changed objects have to be explicitly merged.
A different bug was also dealt with which is that on updating which
seismicity models are available, the main window did not update.