[BUG] Geopoints were not being appended properly sqlalchemy
Due to issues in sqlalchemy, the logic of how the geopoints associated
with hazard point values are now formed into objects in the state
handler instead of at the io stage. This avoids the problem of craeting
a geopoint object when in fact the hazard point should be attached to an
existing geopoint object.

I also think there have been problems with the multiple references to
hazard mdoel run from both the hazardcurve/hazardmap and the
hazardpointvalues objects, but have got around this problem with an
update query that acts directly on the database rather than loading the
objects and updating individually.