[ENH][BUG] Adding realtime functionality and updating simulations
There was a lot of legacy code that didn't really work according to
intended operation. This code has been removed so that fetching data
from web services will now only happen before a forecast, with the
forecast waiting until the data fetch has been completed if required.
There is now real time operation which must be configured in the yml
file as the config update box in the app does not presently work (but is
not high enough priority to spend time on at the moment). The real time
operation also uses the simulator which is a little strange, but that is
how it was origionally created. Really the simulator should be renamed
to something more appropriate as something that can represent the
movement of time for both real time and simulated time.rael time
execution will start as soon as the project is opened in the app. The
start simulation option has not been greyed out, which needs to be done
for real time operation.