[BUG] Fetch data for manual forecast and call old forecasts
Two minor bugs, previously fetch_hydws and fetch_fdsnws were not called
when running a forecast manually if these settings are configured for
the project. Now these methods are run.

Also a condition has been removed so that forecasts that are of status
pending where the forecast time is before the clock time, these are now
run.