[ENH] Validation and bypass of forecast datasources
In the case of natural and induced seismicity, data is sometimes
required, not allowed or optional. In each case, each datasource
should come from one location and project config should be checked so
that the data allowed. Now catalogs and hydraulics data can be added
manually when creating a forecast also.