[BUG] Update cloning behavior when input data not from WS
When the project settings WS urls are empty, either the data has
already been added to a forecast or is not required at all. In this
case, it is safe to use forecast.clone(with_results=True) where the
poorly named with_results just means that input data is also copied
over.