[REFACT] Removing pyqt threading from ramsis
Replacing pyqt classes what were used for threading for the state
handlers from ramsis. It seems to work in much the same way now using
a pure python class. The thread pool and a few other commonly used bits
of code have been moved so as to avoid circular imports.