[ENH] Adding context to logging
The prefect context is used to add detail about the forecast scenario,
model run ids, and the prefect flow run id. There is an issue with all
the logging starting with a unwanted string as the flow is executed
within another flow, but leaving for now as this can be solved with an
update to the ways the flows run in a ater development of ramsis.