[ENH] Making it easier to rerun forecasts
It was very tricky to rerun forecasts before due to idempotency ids
always being used which means the same flow cannot be run more than once
successfully. This is fine in an operational setting, but it has become
clear that there are meny cases where rerunning the same forecast might
be useful. This change means that not adding an idempotency id is
allowed, so the forecast with the same flow run id can be rerun as many times as needed